Move TrieLogProvider setup code to AbstractTrieLogManager Signed-off-by: Simon Dudley <simon.dudley@consensys.net>
