+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/bonsai/cache/CachedWorldStorageManager.java
@@ -21,16 +21,11 @@ import org.hyperledger.besu.ethereum.bonsai.storage.BonsaiWorldStateKeyValueStor
 import org.hyperledger.besu.ethereum.bonsai.storage.BonsaiWorldStateKeyValueStorage.BonsaiStorageSubscriber;
 import org.hyperledger.besu.ethereum.bonsai.storage.BonsaiWorldStateLayerStorage;
 import org.hyperledger.besu.ethereum.bonsai.trielog.AbstractTrieLogManager;
-import org.hyperledger.besu.ethereum.bonsai.trielog.TrieLogFactoryImpl;
 import org.hyperledger.besu.ethereum.bonsai.worldview.BonsaiWorldState;
 import org.hyperledger.besu.ethereum.chain.Blockchain;
 import org.hyperledger.besu.ethereum.core.BlockHeader;
 import org.hyperledger.besu.metrics.ObservableMetricsSystem;
 import org.hyperledger.besu.plugin.BesuContext;
-import org.hyperledger.besu.plugin.services.TrieLogService;
-import org.hyperledger.besu.plugin.services.trielogs.TrieLog;
-import org.hyperledger.besu.plugin.services.trielogs.TrieLogFactory;
-import org.hyperledger.besu.plugin.services.trielogs.TrieLogProvider;
 
 import java.util.ArrayList;
 import java.util.Comparator;
@@ -39,10 +34,7 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.function.Function;
-import java.util.stream.LongStream;
-import java.util.stream.Stream;
 
-import com.google.common.annotations.VisibleForTesting;
 import org.apache.tuweni.bytes.Bytes32;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -226,77 +218,5 @@ public class CachedWorldStorageManager extends AbstractTrieLogManager
   @Override
   public void onCloseStorage() {
     this.cachedWorldStatesByHash.clear();
-  }
-
-  @VisibleForTesting
-  @Override
-  protected TrieLogFactory setupTrieLogFactory(final BesuContext pluginContext) {
-    // if we have a TrieLogService from pluginContext, use it.
-    var trieLogServicez =
-        Optional.ofNullable(pluginContext)
-            .flatMap(context -> context.getService(TrieLogService.class));
-
-    if (trieLogServicez.isPresent()) {
-      var trieLogService = trieLogServicez.get();
-      // push the TrieLogProvider into the TrieLogService
-      trieLogService.configureTrieLogProvider(getTrieLogProvider());
-
-      // configure plugin observers:
-      trieLogService.getObservers().forEach(trieLogObservers::subscribe);
-
-      // return the TrieLogFactory implementation from the TrieLogService
-      return trieLogService.getTrieLogFactory();
-    } else {
-      // Otherwise default to TrieLogFactoryImpl
-      return new TrieLogFactoryImpl();
-    }
-  }
-
-  @VisibleForTesting
-  TrieLogProvider getTrieLogProvider() {
-    return new TrieLogProvider() {
-      @Override
-      public Optional<TrieLog> getTrieLogLayer(final Hash blockHash) {
-        return CachedWorldStorageManager.this.getTrieLogLayer(blockHash);
-      }
-
-      @Override
-      public Optional<TrieLog> getTrieLogLayer(final long blockNumber) {
-        return CachedWorldStorageManager.this
-            .blockchain
-            .getBlockHeader(blockNumber)
-            .map(BlockHeader::getHash)
-            .flatMap(CachedWorldStorageManager.this::getTrieLogLayer);
-      }
-
-      @Override
-      public List<TrieLogRangeTuple> getTrieLogsByRange(
-          final long fromBlockNumber, final long toBlockNumber) {
-        return rangeAsStream(fromBlockNumber, toBlockNumber)
-            .map(blockchain::getBlockHeader)
-            .map(
-                headerOpt ->
-                    headerOpt.flatMap(
-                        header ->
-                            CachedWorldStorageManager.this
-                                .getTrieLogLayer(header.getBlockHash())
-                                .map(
-                                    layer ->
-                                        new TrieLogRangeTuple(
-                                            header.getBlockHash(), header.getNumber(), layer))))
-            .filter(Optional::isPresent)
-            .map(Optional::get)
-            .toList();
-      }
-
-      Stream<Long> rangeAsStream(final long fromBlockNumber, final long toBlockNumber) {
-        if (Math.abs(toBlockNumber - fromBlockNumber) > LOG_RANGE_LIMIT) {
-          throw new IllegalArgumentException("Requested Range too large");
-        }
-        long left = Math.min(fromBlockNumber, toBlockNumber);
-        long right = Math.max(fromBlockNumber, toBlockNumber);
-        return LongStream.range(left, right).boxed();
-      }
-    };
   }
 }
